feat: :zap: Customer Model ID
Id type changed form Long to Integer in Customer Model

@trainingLeader

+++ backend/userdataform/src/main/java/com/mycompany/userdataform/application/UserService.java
@@ -8,9 +8,9 @@ public interface UserService {
 
     Customer saveUser(Customer user);
 
-    Customer updateUser(Customer user, Long userId);
+    Customer updateUser(Customer user, Integer userId);
 
-    void deleteUser(Long userId);
+    void deleteUser(Integer userId);
 
     List<Customer> fetchUsersList();
